# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cytellect private API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/revisions/{rid}/descriptive-preview": {
      "post": {
        "summary": "Descriptive Preview",
        "operationId": "descriptive_preview_v1_revisions__rid__descriptive_preview_post",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DescriptiveRequest"
                  },
                  {
                    "$ref": "#/components/schemas/PagedDescriptiveRequest"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/revisions/{rid}/descriptive": {
      "post": {
        "summary": "Descriptive",
        "operationId": "descriptive_v1_revisions__rid__descriptive_post",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DescriptiveRequest"
                  },
                  {
                    "$ref": "#/components/schemas/PagedDescriptiveRequest"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/revisions/{rid}/common-statistics": {
      "post": {
        "summary": "Calculate",
        "operationId": "calculate_v1_revisions__rid__common_statistics_post",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RegionComparisonRequestV2"
                  },
                  {
                    "$ref": "#/components/schemas/RegionAssociationRequest"
                  }
                ],
                "discriminator": {
                  "propertyName": "mode",
                  "mapping": {
                    "region-experimental-unit": "#/components/schemas/RegionComparisonRequestV2",
                    "region-association": "#/components/schemas/RegionAssociationRequest"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{jid}/common-statistics": {
      "get": {
        "summary": "Result",
        "operationId": "result_v1_jobs__jid__common_statistics_get",
        "parameters": [
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Jid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CommonComparisonView"
                    },
                    {
                      "$ref": "#/components/schemas/AssociationView"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "analysis_kind",
                    "mapping": {
                      "region-comparison": "#/components/schemas/CommonComparisonView",
                      "region-association": "#/components/schemas/AssociationView"
                    }
                  },
                  "title": "Response Result V1 Jobs  Jid  Common Statistics Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssociationView": {
        "properties": {
          "analysis_kind": {
            "type": "string",
            "const": "region-association",
            "title": "Analysis Kind",
            "default": "region-association"
          },
          "source_kind": {
            "type": "string",
            "const": "region-2d",
            "title": "Source Kind",
            "default": "region-2d"
          },
          "region_association_version": {
            "type": "string",
            "const": "1.0.0",
            "title": "Region Association Version",
            "default": "1.0.0"
          },
          "inference_version": {
            "type": "string",
            "title": "Inference Version"
          },
          "revision_id": {
            "type": "string",
            "title": "Revision Id"
          },
          "source_fingerprint": {
            "type": "string",
            "title": "Source Fingerprint"
          },
          "spec": {
            "$ref": "#/components/schemas/RegionAssociationRequest"
          },
          "x_source": {
            "additionalProperties": true,
            "type": "object",
            "title": "X Source"
          },
          "y_source": {
            "additionalProperties": true,
            "type": "object",
            "title": "Y Source"
          },
          "unit_summary": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Unit Summary"
          },
          "unit_ledger": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Unit Ledger"
          },
          "associations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Associations"
          },
          "counts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Counts"
          },
          "missingness": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Missingness"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "figure": {
            "additionalProperties": true,
            "type": "object",
            "title": "Figure"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "inference_version",
          "revision_id",
          "source_fingerprint",
          "spec",
          "x_source",
          "y_source",
          "unit_summary",
          "unit_ledger",
          "associations",
          "counts",
          "missingness",
          "warnings",
          "figure"
        ],
        "title": "AssociationView"
      },
...
      "CommonComparisonView": {
        "properties": {
          "analysis_kind": {
            "type": "string",
            "const": "region-comparison",
            "title": "Analysis Kind",
            "default": "region-comparison"
          },
          "source_kind": {
            "type": "string",
            "const": "region-2d",
            "title": "Source Kind",
            "default": "region-2d"
          },
          "region_comparison_version": {
            "type": "string",
            "const": "2.0.0",
            "title": "Region Comparison Version",
            "default": "2.0.0"
          },
          "inference_version": {
            "type": "string",
            "title": "Inference Version"
          },
          "revision_id": {
            "type": "string",
            "title": "Revision Id"
          },
          "source_fingerprint": {
            "type": "string",
            "title": "Source Fingerprint"
          },
          "spec": {
            "$ref": "#/components/schemas/RegionComparisonRequestV2"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "region": {
            "additionalProperties": true,
            "type": "object",
            "title": "Region"
          },
          "channel": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "source_fields": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Source Fields"
          },
          "source_field_ledger": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Source Field Ledger"
          },
          "observation_ledger": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Observation Ledger"
          },
          "plot_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Plot Data"
          },
          "field_summary": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Field Summary"
          },
          "sample_summary": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sample Summary"
          },
          "unit_summary": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Unit Summary"
          },
          "unit_ledger": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Unit Ledger"
          },
          "pair_ledger": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pair Ledger"
          },
          "counts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Counts"
          },
          "selection": {
            "additionalProperties": true,
            "type": "object",
            "title": "Selection"
          },
          "missingness": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Missingness"
          },
          "excluded_failed_fields": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Excluded Failed Fields"
          },
          "acquisition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Acquisition"
          },
          "comparisons": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Comparisons"
          },
          "means": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Means"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "omnibus": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Omnibus"
          },
          "method_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Method Settings"
          },
          "figure": {
            "additionalProperties": true,
            "type": "object",
            "title": "Figure"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "inference_version",
          "revision_id",
          "source_fingerprint",
          "spec",
          "metric",
          "unit",
          "region",
          "channel",
          "source_fields",
          "source_field_ledger",
          "observation_ledger",
          "plot_data",
          "field_summary",
          "sample_summary",
          "unit_summary",
          "unit_ledger",
          "pair_ledger",
          "counts",
          "selection",
          "missingness",
          "excluded_failed_fields",
          "acquisition",
          "comparisons",
          "means",
          "warnings",
          "omnibus",
          "method_settings",
          "figure"
        ],
        "title": "CommonComparisonView"
      },
...
      "DescriptiveFigurePolicy": {
        "additionalProperties": false,
        "properties": {
          "version": {
            "const": "2.0.0",
            "title": "Version",
            "type": "string"
          },
          "layout": {
            "const": "field-pages",
            "title": "Layout",
            "type": "string"
          }
        },
        "required": [
          "version",
          "layout"
        ],
        "title": "DescriptiveFigurePolicy",
        "type": "object"
      },
...
      "DescriptiveRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "descriptive",
            "title": "Mode"
          },
          "selection": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LegacySelection"
              },
              {
                "$ref": "#/components/schemas/RegionSelection"
              },
              {
                "$ref": "#/components/schemas/NumericalSelection"
              }
            ],
            "title": "Selection",
            "discriminator": {
              "propertyName": "source",
              "mapping": {
                "legacy-cell": "#/components/schemas/LegacySelection",
                "numerical": "#/components/schemas/NumericalSelection",
                "region": "#/components/schemas/RegionSelection"
              }
            }
          },
          "group_by": {
            "type": "string",
            "const": "field",
            "title": "Group By",
            "default": "field"
          },
          "plot": {
            "$ref": "#/components/schemas/DescriptivePlot"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "selection"
        ],
        "title": "DescriptiveRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LegacySelection": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "const": "legacy-cell",
            "title": "Source",
            "type": "string"
          },
          "metric": {
            "enum": [
              "ncl_nucleus_mean",
              "ncl_nucleus_median",
              "ncl_nucleus_integrated",
              "ncl_nucleus_mean_corrected",
              "ncl_nucleus_median_corrected",
              "ncl_nucleus_integrated_corrected",
              "ncl_nucleoli_mean",
              "ncl_nucleoli_median",
              "ncl_nucleoli_integrated",
              "ncl_nucleoli_mean_corrected",
              "ncl_nucleoli_median_corrected",
              "ncl_nucleoli_integrated_corrected",
              "ncl_nucleoplasm_mean",
              "ncl_nucleoplasm_median",
              "ncl_nucleoplasm_integrated",
              "ncl_nucleoplasm_mean_corrected",
              "ncl_nucleoplasm_median_corrected",
              "ncl_nucleoplasm_integrated_corrected",
              "gfp_mean",
              "gfp_median",
              "gfp_integrated",
              "gfp_mean_corrected",
              "gfp_median_corrected",
              "gfp_integrated_corrected",
              "nucleus_area_px",
              "nucleus_area_um2",
              "nucleolar_area_px",
              "nucleolar_area_um2",
              "nucleoplasm_area_px",
              "nucleoplasm_area_um2",
              "nucleolar_count",
              "nucleolar_area_fraction",
              "ncl_nucleoplasm_over_nucleoli",
              "ncl_log2_nucleoplasm_over_nucleoli",
              "ncl_legacy_release"
            ],
            "title": "Metric",
            "type": "string"
          }
        },
        "required": [
          "source",
          "metric"
        ],
        "title": "LegacySelection",
        "type": "object"
      },
...
      "NumericalSelection": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "const": "numerical",
            "title": "Source",
            "type": "string"
          },
          "metric": {
            "const": "value",
            "default": "value",
            "title": "Metric",
            "type": "string"
          }
        },
        "required": [
          "source"
        ],
        "title": "NumericalSelection",
        "type": "object"
      },
...
      "PagedDescriptiveRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "const": "descriptive",
            "title": "Mode",
            "type": "string"
          },
          "selection": {
            "discriminator": {
              "mapping": {
                "legacy-cell": "#/components/schemas/LegacySelection",
                "numerical": "#/components/schemas/NumericalSelection",
                "region": "#/components/schemas/RegionSelection"
              },
              "propertyName": "source"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/LegacySelection"
              },
              {
                "$ref": "#/components/schemas/RegionSelection"
              },
              {
                "$ref": "#/components/schemas/NumericalSelection"
              }
            ],
            "title": "Selection"
          },
          "group_by": {
            "const": "field",
            "default": "field",
            "title": "Group By",
            "type": "string"
          },
          "plot": {
            "$ref": "#/components/schemas/DescriptivePlot"
          },
          "figure_policy": {
            "$ref": "#/components/schemas/DescriptiveFigurePolicy"
          }
        },
        "required": [
          "mode",
          "selection",
          "figure_policy"
        ],
        "title": "PagedDescriptiveRequest",
        "type": "object"
      },
...
      "RegionAssociationRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "region-association",
            "title": "Mode"
          },
          "version": {
            "type": "string",
            "const": "1.0.0",
            "title": "Version"
          },
          "x_selection": {
            "$ref": "#/components/schemas/RegionSelection"
          },
          "y_selection": {
            "$ref": "#/components/schemas/RegionSelection"
          },
          "design": {
            "$ref": "#/components/schemas/ComparisonDesign"
          },
          "conditions": {
            "items": {
              "type": "string",
              "maxLength": 80,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Conditions"
          },
          "acquisition_review": {
            "$ref": "#/components/schemas/AcquisitionReview"
          },
          "missingness_confirmed": {
            "type": "boolean",
            "const": true,
            "title": "Missingness Confirmed"
          },
          "method": {
            "type": "string",
            "enum": [
              "pearson",
              "spearman"
            ],
            "title": "Method"
          },
          "scope": {
            "type": "string",
            "enum": [
              "per-condition",
              "pooled"
            ],
            "title": "Scope",
            "default": "per-condition"
          },
          "pooling_confirmed": {
            "type": "boolean",
            "title": "Pooling Confirmed",
            "default": false
          },
          "aggregation": {
            "type": "string",
            "const": "field-median_sample-mean_unit-mean-v1",
            "title": "Aggregation",
            "default": "field-median_sample-mean_unit-mean-v1"
          },
          "missingness_policy": {
            "type": "string",
            "const": "require-matched-unexcluded-units-v1",
            "title": "Missingness Policy",
            "default": "require-matched-unexcluded-units-v1"
          },
          "plot": {
            "$ref": "#/components/schemas/AssociationPlot"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "version",
          "x_selection",
          "y_selection",
          "design",
          "conditions",
          "acquisition_review",
          "missingness_confirmed",
          "method"
        ],
        "title": "RegionAssociationRequest"
      },
...
      "RegionComparisonRequestV2": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "region-experimental-unit",
            "title": "Mode"
          },
          "version": {
            "type": "string",
            "const": "2.0.0",
            "title": "Version"
          },
          "selection": {
            "$ref": "#/components/schemas/RegionSelection"
          },
          "design": {
            "$ref": "#/components/schemas/ComparisonDesign"
          },
          "conditions": {
            "items": {
              "type": "string",
              "maxLength": 80,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 2,
            "title": "Conditions"
          },
          "comparison_family": {
            "$ref": "#/components/schemas/ComparisonFamily"
          },
          "acquisition_review": {
            "$ref": "#/components/schemas/AcquisitionReview"
          },
          "missingness_confirmed": {
            "type": "boolean",
            "const": true,
            "title": "Missingness Confirmed"
          },
          "aggregation": {
            "type": "string",
            "const": "field-median_sample-mean_unit-mean-v1",
            "title": "Aggregation",
            "default": "field-median_sample-mean_unit-mean-v1"
          },
          "missingness_policy": {
            "type": "string",
            "const": "available-observations_require-unexcluded-units-v1",
            "title": "Missingness Policy",
            "default": "available-observations_require-unexcluded-units-v1"
          },
          "plot": {
            "$ref": "#/components/schemas/CommonComparisonPlot"
          },
          "test": {
            "type": "string",
            "enum": [
              "welch-t",
              "paired-t",
              "mann-whitney-u",
              "wilcoxon"
            ],
            "title": "Test"
          },
          "omnibus": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "welch-anova",
                  "kruskal-wallis"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Omnibus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "version",
          "selection",
          "design",
          "conditions",
          "comparison_family",
          "acquisition_review",
          "missingness_confirmed",
          "test"
        ],
        "title": "RegionComparisonRequestV2"
      },
...
      "RegionSelection": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "const": "region",
            "title": "Source",
            "type": "string"
          },
          "region_set_id": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Region Set Id",
            "type": "string"
          },
          "channel_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Channel Id"
          },
          "metric": {
            "enum": [
              "area_px",
              "area_um2",
              "mean",
              "median",
              "integrated",
              "mean_corrected",
              "median_corrected",
              "integrated_corrected"
            ],
            "title": "Metric",
            "type": "string"
          }
        },
        "required": [
          "source",
          "region_set_id",
          "metric"
        ],
        "title": "RegionSelection",
        "type": "object"
      },
...
    }
  }
}